Fix: Create repo if it does not exist

+++ openapi_server/controllers/default_controller.py
@@ -899,10 +899,16 @@ async def run_pipeline(
             repo_branch
         )
     else:
+        _create_repo = False
         _repo = gh_utils.get_repository(pipeline_repo)
-        if not gh_utils.get_repo_content(pipeline_repo):
-            # Re-create with content (i.e. README)
-            _repo.delete()
+        if not _repo:
+            _create_repo = True
+        else:
+            if not gh_utils.get_repo_content(pipeline_repo):
+                _repo.delete()
+                _create_repo = True
+        if _create_repo:
+            # Create repo with content (i.e. README)
             _repo = gh_utils.create_org_repository(pipeline_repo)
         pipeline_repo_branch = _repo.default_branch
